1. make ServiceEntity an abstract class DefaultServiceEntity 2. remove redundant lock in checkAllForReconnect and add elaboration on mNeedReconnect instance var
